* fix/pr74-g03-points-overflow:
  fix(g03): extend ::int→::bigint to cron audit scripts (3 files)
  fix(g03): SUM(amount)::int -> ::bigint to prevent overflow, 4-end sync

+++ fengyu-admin/src/cron/steps/audit-points-balance.ts
@@ -24,7 +24,7 @@ export interface PointsAuditResult {
 export async function auditPointsBalance(db: Db): Promise<PointsAuditResult> {
   const rows = (await db.execute(sql`
     WITH sums AS (
-      SELECT user_id, COALESCE(SUM(amount), 0)::int AS total_from_txns
+      SELECT user_id, COALESCE(SUM(amount), 0)::bigint AS total_from_txns
       FROM point_transactions
       GROUP BY user_id
     )
