+++ fengyu-admin/src/lib/points-settle.ts
@@ -89,7 +89,7 @@ export async function settlePointsForOrder(
 
   // 4. 已发积分合计（按 ref_order_id 聚合，等级升级/兑换等非本链流水自然排除）
   const grantedRes = await tx.execute(sql`
-    SELECT COALESCE(SUM(amount), 0)::int AS granted
+    SELECT COALESCE(SUM(amount), 0)::bigint AS granted
       FROM point_transactions
      WHERE ref_order_id = ${originalSaleOrderId}
   `)
